+++ open-entities-lib/src/components/mod.rs
@@ -1,15 +1,23 @@
 pub mod base_move_speed;
 pub mod entity_type;
 pub mod faction;
+pub mod group;
 pub mod health;
+pub mod manual_active;
+pub mod member_of;
 pub mod move_target;
+pub mod order_source;
 pub mod position;
 pub mod velocity;
 
 pub use base_move_speed::BaseMoveSpeed;
 pub use entity_type::EntityType;
 pub use faction::Faction;
+pub use group::Group;
 pub use health::Health;
+pub use manual_active::ManualActive;
+pub use member_of::MemberOf;
 pub use move_target::MoveTarget;
+pub use order_source::OrderSource;
 pub use position::Position;
 pub use velocity::Velocity;